Restructure: QA
- Move the files from `/Management/User` to `/Users`.
