+++ app/Kwality.UVault.QA/Users/Mappers/Auth0.User.Update.Operation.Mapper.Tests.cs
@@ -22,22 +22,24 @@
 // =                FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 // =                OTHER DEALINGS IN THE SOFTWARE.
 // =====================================================================================================================
-namespace Kwality.UVault.QA.Management.User.Mappers;
+namespace Kwality.UVault.QA.Users.Mappers;
 
 using AutoFixture.Xunit2;
 
 using FluentAssertions;
 
+using global::Auth0.ManagementApi.Models;
+
 using JetBrains.Annotations;
 
+using Kwality.UVault.Auth0.Users.Operations.Mappers;
 using Kwality.UVault.QA.Internal.Xunit.Traits;
 using Kwality.UVault.Users.Exceptions;
-using Kwality.UVault.Users.Operations.Mappers;
 using Kwality.UVault.Users.Operations.Mappers.Abstractions;
 
 using Xunit;
 
-public sealed class UserCreateOperationMapperTests
+public sealed class Auth0UserUpdateOperationMapperTests
 {
     [UserManagement]
     [AutoData]
@@ -45,15 +47,15 @@ public sealed class UserCreateOperationMapperTests
     internal void MapSourceToDestination_InvalidDestination_RaisesException(ModelOne model)
     {
         // ARRANGE.
-        var mapper = new UserCreateOperationMapper();
+        var mapper = new UserUpdateOperationMapper();
 
         // ACT.
         Action act = () => mapper.Create<ModelOne, ModelTwo>(model);
 
         // ASSERT.
         act.Should()
-           .Throw<UserCreationException>()
-           .WithMessage($"Invalid {nameof(IUserOperationMapper)}: Destination is NOT `{nameof(ModelOne)}`.");
+           .Throw<UserUpdateException>()
+           .WithMessage($"Invalid {nameof(IUserOperationMapper)}: Destination is NOT `{nameof(UserUpdateRequest)}`.");
     }
 
     [UserManagement]
@@ -62,14 +64,22 @@ public sealed class UserCreateOperationMapperTests
     internal void MapSourceToDestination_Succeeds(ModelOne model)
     {
         // ARRANGE.
-        var mapper = new UserCreateOperationMapper();
+        var mapper = new UserUpdateOperationMapper();
 
         // ACT.
-        ModelOne result = mapper.Create<ModelOne, ModelOne>(model);
+        UserUpdateRequest result = mapper.Create<ModelOne, UserUpdateRequest>(model);
 
         // ASSERT.
         result.Should()
-              .BeEquivalentTo(model);
+              .BeEquivalentTo(new UserUpdateRequest());
+    }
+
+    private sealed class UserUpdateOperationMapper : Auth0UserUpdateOperationMapper
+    {
+        protected override UserUpdateRequest Map<TSource>(TSource source)
+        {
+            return new UserUpdateRequest();
+        }
     }
 
     [UsedImplicitly]
