+++ app/Kwality.UVault.QA/Users/User.Management.Default.Tests.cs
@@ -22,7 +22,7 @@
 // =                FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 // =                OTHER DEALINGS IN THE SOFTWARE.
 // =====================================================================================================================
-namespace Kwality.UVault.QA.Management.User;
+namespace Kwality.UVault.QA.Users;
 
 using AutoFixture;
 using AutoFixture.Kernel;
